Optimize table cleanup
When creating foreign keys, referred tables should be created before
referring tables. This means, that `@transient_tables` will have
referred tables come before that which are referring to them. While
making a first iteration over the list to remove them, the loop will
fail and push the referred tables to the end, to be able to remove them
at the next iteration.

But if we traverse the list in the reverse order, we improving the
likelihood of being able to remove all the tables on the first iteration
and thus reducing the number of database queries.